# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexora Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Content {
  id: number;
  title: string;
  description: string;
  /** movie, series, documentary, short */
  type: string;
  genre: string;
  year: number;
  rating: number;
  duration: string;
  /** @nullable */
  episodeCount?: number | null;
  /** @nullable */
  seasons?: number | null;
  thumbnailUrl: string;
  backdropUrl: string;
  /** @nullable */
  trailerUrl?: string | null;
  isFeatured: boolean;
  isTrending: boolean;
  isNewRelease: boolean;
  cast: string[];
  /** @nullable */
  director?: string | null;
}

export interface Genre {
  id: number;
  name: string;
  slug: string;
}

export interface SearchResult {
  items: Content[];
  total: number;
}

export interface WatchlistItem {
  id: number;
  contentId: number;
  addedAt: string;
  content: Content;
}

export interface WatchlistInput {
  contentId: number;
}

export type ListContentParams = {
genre?: string;
type?: string;
limit?: number;
offset?: number;
};

export type SearchContentParams = {
q: string;
};