# lib/api-zod/src/generated/types/watchlistItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexora Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Content } from './content';

export interface WatchlistItem {
  id: number;
  contentId: number;
  addedAt: Date;
  content: Content;
}